sweet-crud/requirements.txt
Django>=4.0,<5.0
djangorestframework-simplejwt django-cors-headers
